refactor: Expose k8s package
- Move internal/k8s/ to k8s/ to make it public. This allows external
  projects to import the package to provision Cilium programmatically.
- Add a package level documentation to clarify that the APIs might
  change before cilium-cli hits v1.0.0.
- Update CODEOWNERS.

Fixes #618, fixes #776
