+++ connectivity/check/check.go
@@ -13,7 +13,7 @@ import (
 	"github.com/cilium/cilium/api/v1/observer"
 
 	"github.com/cilium/cilium-cli/connectivity/filters"
-	"github.com/cilium/cilium-cli/internal/k8s"
+	"github.com/cilium/cilium-cli/k8s"
 )
 
 type Parameters struct {
