+++ install/node_init.go
@@ -7,8 +7,8 @@ import (
 	"github.com/cilium/cilium/pkg/versioncheck"
 	appsv1 "k8s.io/api/apps/v1"
 
-	"github.com/cilium/cilium-cli/internal/k8s"
 	"github.com/cilium/cilium-cli/internal/utils"
+	"github.com/cilium/cilium-cli/k8s"
 )
 
 func needsNodeInit(k k8s.Kind) bool {
