+++ k8s/doc.go
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2018 Authors of Cilium
+
+// Package k8s provides various helper functions for interacting with Kubernetes
+// APIs.
+//
+// IMPORTANT: Please note that the APIs of this package might change without any
+// notice until cilium-cli reaches v1.0.0.
+package k8s
